pyproject.toml
[tool.poetry]
name = "read-cccd-service"
version = "0.1.0"
description = ""
authors = ["toandev020101 <toandev020101@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
easyocr = "^1.7.1"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pillow = "^10.3.0"
gunicorn = "^22.0.0"
pyzbar = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
